requirements.txt
sqlparse ~= 0.2.4
django ~= 2.2.5
djangorestframework ~= 3.10.2
djongo ~= 1.2.31
django-cors-headers ~= 3.1.0
requests ~= 2.22.0
bs4 ~= 0.0.1
Markdown ~= 3.1.1
ruamel.yaml ~= 0.16.5
ruamel.yaml.clib ~= 0.1.2
html2markdown ~= 0.1.7
unidecode ~= 1.1.1
